Update README.md
If you have been waiting to find a way to backup IMAP on a Mac, you are at the right place. Here you are going to find the best IMAP backup tool for mac and windows. Most email users are not able to find a reliable tool because there is a lot of confusion among email users regarding the perfect way to manage emails.